+++ pkg/interface.go
@@ -1,7 +1,10 @@
 package bhlquest
 
+import "github.com/gnames/bhlquest/pkg/ent/answer"
+
 // BHLQuest provides functionality needed to apply AI to BHL.
 type BHLQuest interface {
 	// Init bootstraps AI engines providing necessary data and metadata.
 	Init() error
+	Ask(q string) (answer.Answer, error)
 }